pos/backend/requirements.txt
# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt,argon2]>=1.7.4
argon2-cffi>=21.3.0

# Environment and Configuration
python-dotenv>=1.0.0

# Data Validation and Serialization
pydantic>=2.0.0
email-validator>=2.1.0

# HTTP client for ERP integration
httpx>=0.25.0

# Date and time handling
python-dateutil>=2.8.2